UX
A copy of the selected piece is picked up on your mouse when holding left mouse down on a piece, it is colored red for invalid moves, green for valid moves.
While holding a piece, mousing over a valid attack target will turn the target red, and a valid defend target will turn green.