bug fixes
Changing your display name is now automatically updated over the network (as are typos)
Removed update name button
If timer/clock is off, the end game panel will no longer display the length of the game (but it's still tracked internally)
Fixed a bug where B1 was always highlighted as blue when loaded a completed game. Now it will only highlight the last play if the game was saved in an uncompleted state.