bug fix
when hosting a game, the preview move toggle was being defaulted to on after the event bind that sends a message to the client when changed. By default prior to the event bind, we are able to ensure that the message isn't sent unless the player clicks on the toggle.

(We might have a bug if the host toggles off before a client connects???)