Feature work
Added text when it becomes your turn in a multiplayer game
move previews now include hexes with pieces that are guarded by the hovered piece
Prevented the player from pressing "Connect" multiple times in a row if there is a high latency to the host. It now waits for a connection failure before allowing another connect attempt.