feature work
Jails are now sorted
Fixed a bug where check was trying to move the king to nearby locations with allied pieces on them, causing a dictionary conflict due to a duplicate key.
in sandbox mode, the camera flips when the turn changes.
In sandbox mode, the keys flip when the camera flips.
Added value to all pieces, this is used for sorting in the jails, but may also be used for the AI in the future to award different points when pieces are taken.