Added free place mode
while in sandbox mode, clicking free place mode at the bottom will allow pieces to be placed anywhere, including in the jail, as well as being pulled out of the jail.
Added a Kings Only button while in free place mode that removes everything from the board that isn't a king.