Clean up CheckCompleteMessage
When receiving a whole message followed by a second incomplete message,
readBufferStart would be set to the length of the second message when
the second message is eventually read. Instead it should've been set to
the length of the first message + the length of the second message. This
bug was hard to spot because of the "pos" variable. Removing the pos
variable makes the code easier to read, and eliminates the bug.