Introduce StandaloneRules assembly
To improve HexCoven, the matchmaking server, I would like to introduce
two features: An AI opponent, and move validation. I do not want to
reinvent the hexachessagon wheel, so it would be nice if I could re-use
existing game logic. StandaloneRules is the assembly where I intend to
move that relevant game logic.