Bug fixes and feature work!
ctrl, alt, shift now modify the color of arrows when drawn with the handicap overlay turned off.
Fixed several bugs surrounding the lastTurn text on the top right and how they interacted with free place mode.
Fixed a bug where a hex would sometimes not be deselected properly when cancelling a move by dropping on top of the jails.
Fixed a visual bug where the white pawn's move speed got set to 0 somehow, making it appear that the white pawns couldn't be moved, even though the boardstate was being updated correctly.