bug fixes + feature
added tooltip that tells the name of hovered pieces while the handicap overlay is enabled.
fixed a bug where the turn number was listed wrong in the turn panel when black caused a draw
Fixed a bug where the turn number listed in the turn history panel was wrong when white caused a draw
fixed a bug where the turn number listed in the turn panel was wrong when accepting a draw offer.
Fixed a bug that could freeze the game when entering an unsupported emoji into the name input field
fixed a bug when reviewing a turn to the turn before a promotion happened caused the pawn to move from it's starting loc in a default board state rather than the location of the promoted piece