Work on lobby
It is now possible to start an AI game from the versus menu. AI settings WIP
Fixed a bug where changing your name as a client updated both your name and your opponent's name to be yours.
Removed Lobby's dependency on networker, it is now set in a mode.
Clarified which team you will be playing as in the lobby.