Bug fix
fixed a bug when using freeplace mode to illegally capture a white piece with black which would cause the last move calculation to discover a white piece leaving the board before finding a black move and jumping to a conclusion that the white piece was jailed and that it would be white's turn, but that's not true. Now we only check the team who just played for it's moves instead of all pieces.